requirements.txt
click>=8.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
rich>=13.0.0
langchain>=0.1.0
openai>=1.0.0
gitpython>=3.1.0